+++ custom_components/wattpilot/const.py
@@ -17,7 +17,6 @@ CONF_CLOUD_API: Final = "cloud_api"
 CONF_CLOUD: Final = "cloud"
 CONF_CONNECTION: Final = "connection"
 CONF_LOCAL: Final = "local"
-CONF_PUSH_ENTITIES: Final = "push_entities"
 CONF_SERIAL: Final = "serial"
 
 DEFAULT_TIMEOUT: Final = 15